+++ integration/spark-common/src/main/scala/org/apache/spark/sql/hive/DistributionUtil.scala
@@ -29,6 +29,7 @@ import org.apache.carbondata.core.carbon.datastore.block.Distributable
 import org.apache.carbondata.core.constants.CarbonCommonConstants
 import org.apache.carbondata.core.util.CarbonProperties
 import org.apache.carbondata.spark.load.CarbonLoaderUtil
+import org.apache.carbondata.spark.rdd.SparkCommonEnv
 
 object DistributionUtil {
   @transient
@@ -215,9 +216,6 @@ object DistributionUtil {
     nodes.distinct.toSeq
   }
 
-  // Hack for spark2 integration
-  var numExistingExecutors: Int = _
-
   /**
    * Requesting the extra executors other than the existing ones.
    *
@@ -233,13 +231,11 @@ object DistributionUtil {
       hostToLocalTaskCount: Map[String, Int] = Map.empty): Boolean = {
     sc.schedulerBackend match {
       case b: CoarseGrainedSchedulerBackend =>
-        LOGGER
-          .info(
+        LOGGER.info(
             s"number of required executors are = $requiredExecutors and existing executors are = " +
-            s"$numExistingExecutors")
+            s"${SparkCommonEnv.numExistingExecutors}")
         if (requiredExecutors > 0) {
-          LOGGER
-            .info(s"Requesting total executors: $requiredExecutors")
+          LOGGER.info(s"Requesting total executors: $requiredExecutors")
           b.requestTotalExecutors(requiredExecutors, localityAwareTasks, hostToLocalTaskCount)
         }
         true
